WWT.Providers: deprecate unused/broken APIs
Closes #204.
