--- src/WWT.Providers/TileProviders/EarthMerBathProvider.cs
@@ -1,79 +1,0 @@
-#nullable disable
-
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-using WWTWebservices;
-
-namespace WWT.Providers
-{
-    [RequestEndpoint("/wwtweb/EarthMerBath.aspx")]
-    public class EarthMerBathProvider : RequestProvider
-    {
-        private readonly IPlateTilePyramid _plateTiles;
-        private readonly WwtOptions _options;
-        private readonly IVirtualEarthDownloader _veDownloader;
-
-        public EarthMerBathProvider(IPlateTilePyramid plateTiles, WwtOptions options, IVirtualEarthDownloader veDownloader)
-        {
-            _plateTiles = plateTiles;
-            _options = options;
-            _veDownloader = veDownloader;
-        }
-
-        public override string ContentType => ContentTypes.Png;
-
-        public override async Task RunAsync(IWwtContext context, CancellationToken token)
-        {
-            (var errored, var level, var tileX, var tileY) = await HandleLXYQParameter(context, token);
-            if (errored)
-                return;
-
-            if (level > 20)
-            {
-                await context.Response.WriteAsync("No image", token);
-                context.Response.Close();
-            }
-            else if (level < 8)
-            {
-                context.Response.ContentType = "image/png";
-
-                using (Stream s = await _plateTiles.GetStreamAsync(_options.WwtTilesDir, "BmngMerBase.plate", level, tileX, tileY, token))
-                {
-                    await s.CopyToAsync(context.Response.OutputStream, token);
-                    context.Response.Flush();
-                    context.Response.End();
-                }
-            }
-            else if (level < 10)
-            {
-                int L = level;
-                int X = tileX;
-                int Y = tileY;
-                int powLev5Diff = (int)Math.Pow(2, L - 2);
-                int X32 = X / powLev5Diff;
-                int Y32 = Y / powLev5Diff;
-
-                int L5 = L - 2;
-                int X5 = X % powLev5Diff;
-                int Y5 = Y % powLev5Diff;
-
-                context.Response.ContentType = "image/png";
-
-                using (Stream s = await _plateTiles.GetStreamAsync(_options.WwtTilesDir, $"BmngMerL2X{X32}Y{Y32}.plate", L5, X5, Y5, token))
-                {
-                    await s.CopyToAsync(context.Response.OutputStream, token);
-                    context.Response.Flush();
-                    context.Response.End();
-                }
-            }
-            else
-            {
-                using var veTile = await _veDownloader.DownloadVeTileAsync(VirtualEarthTile.Ortho, level, tileX, tileY, token);
-
-                await veTile.CopyToAsync(context.Response.OutputStream, token);
-            }
-        }
-    }
-}
