--- src/WWT.Providers/TileProviders/Hirise.Aspx.cs
@@ -1,14 +1,0 @@
-#nullable disable
-
-using WWTWebservices;
-
-namespace WWT.Providers
-{
-    public abstract partial class HiRise : RequestProvider
-    {
-        public uint ComputeHash(int level, int x, int y)
-        {
-            return DirectoryEntry.ComputeHash(level + 128, x, y);
-        }
-    }
-}
