--- tests/WWT.Providers.Tests/Demmarsnewprovidertests.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Threading.Tasks;
-using WWTWebservices;
-using Xunit;
-
-namespace WWT.Providers.Tests
-{
-    public class DemMarsNewProviderTests : ProviderTests<DemMarsNewProvider>
-    {
-        protected override int MaxLevel => 17;
-
-        protected override Action<IResponse> StreamExceptionResponseHandler => null;
-
-        protected override void ExpectedResponseAboveMaxLevel(IResponse response)
-        {
-            Assert.Empty(response.OutputStream.ToArray());
-        }
-
-        protected override Task<Stream> GetStreamFromPlateTilePyramidAsync(IPlateTilePyramid plateTiles, int level, int x, int y)
-        {
-            var index = DirectoryEntry.ComputeHash(level + 128, x, y) % 400;
-
-            return plateTiles.GetStreamAsync(@"\\wwt-mars\marsroot\dem\", $"marsToastDem_{index}.plate", -1, level, x, y, default);
-        }
-    }
-}
