--- tests/WWT.Providers.Tests/Rasstoastprovidertests.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Threading.Tasks;
-using WWTWebservices;
-using Xunit;
-
-namespace WWT.Providers.Tests
-{
-    public class RassToastProviderTests:ProviderTests<RassToastProvider>
-    {
-        protected override int MaxLevel => 7;
-
-        protected override Action<IResponse> NullStreamResponseHandler => null;
-
-        protected override Action<IResponse> StreamExceptionResponseHandler => null;
-
-
-        protected override Task<Stream> GetStreamFromPlateTilePyramidAsync(IPlateTilePyramid plateTiles, int level, int x, int y)
-            => plateTiles.GetStreamAsync(Options.WwtTilesDir, "RASS.plate", level, x, y, default);
-
-        protected override void ExpectedResponseAboveMaxLevel(IResponse response)
-        {
-            Assert.Empty(response.ContentType);
-            Assert.Empty(response.OutputStream.ToArray());
-        }
-    }
-}
